+++ ipo-service-provider/src/main/java/com/yrdce/ipo/modules/warehouse/entity/IpoWarehouseStock.java
@@ -1,17 +1,19 @@
 package com.yrdce.ipo.modules.warehouse.entity;
 
+import java.math.BigDecimal;
+
 public class IpoWarehouseStock {
 	private String stockid;
 
 	private String commodityid;
 
-	private Long storagenum;
+	private BigDecimal storagenum;
 
-	private Long forzennum;
+	private BigDecimal forzennum;
 
-	private Long availablenum;
+	private BigDecimal availablenum;
 
-	private Long outboundnum;
+	private BigDecimal outboundnum;
 
 	private Long warehouseid;
 
@@ -71,35 +73,35 @@ public class IpoWarehouseStock {
 		this.commodityid = commodityid == null ? null : commodityid.trim();
 	}
 
-	public Long getStoragenum() {
+	public BigDecimal getStoragenum() {
 		return storagenum;
 	}
 
-	public void setStoragenum(Long storagenum) {
+	public void setStoragenum(BigDecimal storagenum) {
 		this.storagenum = storagenum;
 	}
 
-	public Long getForzennum() {
+	public BigDecimal getForzennum() {
 		return forzennum;
 	}
 
-	public void setForzennum(Long forzennum) {
+	public void setForzennum(BigDecimal forzennum) {
 		this.forzennum = forzennum;
 	}
 
-	public Long getAvailablenum() {
+	public BigDecimal getAvailablenum() {
 		return availablenum;
 	}
 
-	public void setAvailablenum(Long availablenum) {
+	public void setAvailablenum(BigDecimal availablenum) {
 		this.availablenum = availablenum;
 	}
 
-	public Long getOutboundnum() {
+	public BigDecimal getOutboundnum() {
 		return outboundnum;
 	}
 
-	public void setOutboundnum(Long outboundnum) {
+	public void setOutboundnum(BigDecimal outboundnum) {
 		this.outboundnum = outboundnum;
 	}
